config: HCL2-flavored detection of variable references
The existing function DetectVariables uses the HIL API to find
interpolation sequences and extract variable references from them. That's
not a suitable path for configs coming from the HCL2 config loader, since
it ends up holding a HCL2 AST instead.

This new function uses the HCL2 API to detect references (which are called
"traversals" in HCL2 terminology) and produce similar-but-not-identical
InterpolatedVariable instances for them.

The resulting slice is not suitable for passing to the existing core code
that constructs the HIL scope because it captures only a subset of the
information that DetectVariables would capture. Thus a subsequent commit
will introduce a new scope builder designed for HCL2 that is able to
consume the results of DetectVariablesHCL2 and produce a native HCL2
variable table.